Versão final que foi entregue ao professor. Coeficiente soften finalizado Corrigido calculo da refração

+++ RayTracing/Src/Light.h
@@ -25,7 +25,7 @@ public:
 	Light( void );
 	Light( const Vector3D &aPosition, const ColorRGBf &aColor = ColorRGBf(1,1,1));
 
-	virtual double evaluateIlumination( const Vector3D& aPoint, const Scene* aScene, double aTiny) = 0;
+	virtual float evaluateIlumination( const Vector3D& aPoint, const Scene* aScene, double aTiny) = 0;
 
 	inline const ColorRGBf &getColor()
 	{
